requirements.txt
tiktoken
asyncio==3.4.3
pybase64
langchain==0.0.246
openai==0.27.4
faiss-gpu
faiss-cpu